+++ packages/component_code_gen/code_gen/transform_code.py
@@ -2,7 +2,7 @@ from dotenv import load_dotenv
 load_dotenv()
 
 import openai
-import templates.transform as templates
+import templates.transform_to_action as templates
 from config.config import config
