+++ packages/component_code_gen/templates/generate_actions.py
